# docs/api/openapi.yml
components:
  schemas:
...
    ErrorsJSON:
      properties:
        code:
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorsResponseerroradditional'
          type: array
        message:
          type: string
        url:
          type: string
      required:
      - code
      - message
      type: object
    ErrorsResponseerroradditional:
      properties:
        message:
          type: string
      type: object
...
    RenderErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorsJSON'
        status:
          type: string
      required:
      - status
      - error
      type: object
    ServiceaccounttypesFactorAPIKey:
      properties:
        createdAt:
          format: date-time
          type: string
        expiresAt:
          minimum: 0
          type: integer
        id:
          type: string
        key:
          type: string
        lastObservedAt:
          format: date-time
          type: string
        name:
          type: string
        serviceAccountId:
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - id
      - key
      - expiresAt
      - lastObservedAt
      - serviceAccountId
      type: object
    ServiceaccounttypesGettableFactorAPIKeyWithKey:
      properties:
        id:
          type: string
        key:
          type: string
      required:
      - id
      - key
      type: object
    ServiceaccounttypesPostableFactorAPIKey:
      properties:
        expiresAt:
          minimum: 0
          type: integer
        name:
          type: string
      required:
      - name
      - expiresAt
      type: object
    ServiceaccounttypesPostableServiceAccount:
      properties:
        name:
          type: string
        roles:
          items:
            $ref: '#/components/schemas/ServiceaccounttypesPostableServiceAccountRole'
          type: array
      required:
      - name
      - roles
      type: object
    ServiceaccounttypesPostableServiceAccountRole:
      properties:
        name:
          type: string
      required:
      - name
      type: object
    ServiceaccounttypesServiceAccountRole:
      properties:
        createdAt:
          format: date-time
          type: string
        id:
          type: string
        name:
          type: string
        roleId:
          type: string
        serviceAccountId:
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - id
      - serviceAccountId
      - roleId
      - name
      type: object
    ServiceaccounttypesServiceAccountWithRoles:
      properties:
        createdAt:
          format: date-time
          type: string
        email:
          type: string
        id:
          type: string
        name:
          type: string
        orgId:
          type: string
        roles:
          items:
            $ref: '#/components/schemas/ServiceaccounttypesServiceAccountRole'
          type: array
        status:
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - id
      - name
      - email
      - status
      - orgId
      - roles
      type: object
    ServiceaccounttypesUpdatableFactorAPIKey:
      properties:
        expiresAt:
          minimum: 0
          type: integer
        name:
          type: string
      required:
      - name
      - expiresAt
      type: object
    ServiceaccounttypesUpdatableServiceAccount:
      properties:
        name:
          type: string
        roles:
          items:
            $ref: '#/components/schemas/ServiceaccounttypesPostableServiceAccountRole'
          type: array
      required:
      - name
      - roles
      type: object
    ServiceaccounttypesUpdatableServiceAccountStatus:
      properties:
        status:
          type: string
      required:
      - status
      type: object
...
    TypesIdentifiable:
      properties:
        id:
          type: string
      required:
      - id
      type: object
...
  securitySchemes:
    api_key:
      description: API Keys
      in: header
      name: SigNoz-Api-Key
      type: apiKey
    tokenizer:
      bearerFormat: Tokenizer
      description: Tokens generated by the tokenizer
      scheme: bearer
      type: http
info:
  description: OpenTelemetry-Native Logs, Metrics and Traces in a single pane
  title: SigNoz
  version: ""
openapi: 3.0.3
paths:
...
  /api/v1/service_accounts:
    get:
      deprecated: false
      description: This endpoint lists the service accounts for an organisation
      operationId: ListServiceAccounts
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  data:
                    items:
                      $ref: '#/components/schemas/ServiceaccounttypesServiceAccountWithRoles'
                    type: array
                  status:
                    type: string
                required:
                - status
                - data
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: List service accounts
      tags:
      - serviceaccount
    post:
      deprecated: false
      description: This endpoint creates a service account
      operationId: CreateServiceAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceaccounttypesPostableServiceAccount'
      responses:
        "201":
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/TypesIdentifiable'
                  status:
                    type: string
                required:
                - status
                - data
                type: object
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Create service account
      tags:
      - serviceaccount
  /api/v1/service_accounts/{id}:
    delete:
      deprecated: false
      description: This endpoint deletes an existing service account
      operationId: DeleteServiceAccount
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          content:
            application/json:
              schema:
                type: string
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Deletes a service account
      tags:
      - serviceaccount
    get:
      deprecated: false
      description: This endpoint gets an existing service account
      operationId: GetServiceAccount
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ServiceaccounttypesServiceAccountWithRoles'
                  status:
                    type: string
                required:
                - status
                - data
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Gets a service account
      tags:
      - serviceaccount
    put:
      deprecated: false
      description: This endpoint updates an existing service account
      operationId: UpdateServiceAccount
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceaccounttypesUpdatableServiceAccount'
      responses:
        "204":
          content:
            application/json:
              schema:
                type: string
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Updates a service account
      tags:
      - serviceaccount
  /api/v1/service_accounts/{id}/keys:
    get:
      deprecated: false
      description: This endpoint lists the service account keys
      operationId: ListServiceAccountKeys
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  data:
                    items:
                      $ref: '#/components/schemas/ServiceaccounttypesFactorAPIKey'
                    type: array
                  status:
                    type: string
                required:
                - status
                - data
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: List service account keys
      tags:
      - serviceaccount
    post:
      deprecated: false
      description: This endpoint creates a service account key
      operationId: CreateServiceAccountKey
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceaccounttypesPostableFactorAPIKey'
      responses:
        "201":
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ServiceaccounttypesGettableFactorAPIKeyWithKey'
                  status:
                    type: string
                required:
                - status
                - data
                type: object
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Create a service account key
      tags:
      - serviceaccount
  /api/v1/service_accounts/{id}/keys/{fid}:
    delete:
      deprecated: false
      description: This endpoint revokes an existing service account key
      operationId: RevokeServiceAccountKey
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      - in: path
        name: fid
        required: true
        schema:
          type: string
      responses:
        "204":
          content:
            application/json:
              schema:
                type: string
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Revoke a service account key
      tags:
      - serviceaccount
    put:
      deprecated: false
      description: This endpoint updates an existing service account key
      operationId: UpdateServiceAccountKey
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      - in: path
        name: fid
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceaccounttypesUpdatableFactorAPIKey'
      responses:
        "204":
          content:
            application/json:
              schema:
                type: string
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Updates a service account key
      tags:
      - serviceaccount
  /api/v1/service_accounts/{id}/status:
    put:
      deprecated: false
      description: This endpoint updates an existing service account status
      operationId: UpdateServiceAccountStatus
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceaccounttypesUpdatableServiceAccountStatus'
      responses:
        "204":
          content:
            application/json:
              schema:
                type: string
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Updates a service account status
      tags:
      - serviceaccount
...